requirements.txt
fastapi[all]
sqlalchemy
passlib
python-jose
python-dotenv
bcrypt